requirements.txt
fastapi
uvicorn[standard]
pydantic
transformers
gtts
fpdf
markdown
beautifulsoup4
duckduckgo-search
pytrends
pillow
typer
